requirements.txt
bcrypt==4.2.1
certifi==2024.12.14
click==8.1.8
dnspython==2.3.0
Flask==2.2.5
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Flask-Session>=0.4
gunicorn==23.0.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.5
MarkupSafe==2.1.5
packaging==24.0
PyJWT==2.8.0
pymongo==4.7.3
python-dotenv==0.21.1
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
